requirements.txt
django>5.0
djangorestframework
django_browser_reload
drf-spectacular
WeasyPrint
python-decouple
dj-database-url
psycopg2
psycopg2-binary
django-debug-toolbar
django_browser_reload
django-extensions
celery
django-celery
pybloom-live
gunicorn
markdown


